Separated upload logo to different component
